+++ Tests/DevFoundationTests/Testing Helpers/DefaultRemoteContentBundleActor.swift
@@ -1,0 +1,13 @@
+//
+//  DefaultRemoteContentBundleActor.swift
+//  DevFoundation
+//
+//  Created by Prachi Gauriar on 1/13/26.
+//
+
+import Foundation
+
+@globalActor
+actor DefaultRemoteContentBundleActor {
+    static let shared = DefaultRemoteContentBundleActor()
+}